requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# AI Model - Ollama
ollama==0.3.0

# Vector Database & Embeddings
qdrant-client==1.7.0
FlagEmbedding==1.2.10

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pdf2image==1.17.0
Pillow==10.4.0

# Database
aiosqlite==0.20.0

# Authentication & Security
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Session Management
itsdangerous==2.2.0

# Environment Variables
python-dotenv==1.0.1

# Template Engine
jinja2==3.1.4

# HTTP Client
httpx==0.27.0

# Metrics & Monitoring
prometheus-client==0.20.0

# Utilities
python-multipart==0.0.9
aiofiles==24.1.0

# Testing
reportlab==4.2.0
